# openapi/v1.yaml
openapi: 3.1.0
info:
  title: registerwatch
  description: Gambling regulators' public registers, one schema per register, with history.
  contact:
    name: registerwatch
    url: https://github.com/mishmish004/RegisterWatch
  license:
    name: All rights reserved
    url: https://choosealicense.com/no-permission/
  version: 0.2.0
servers:
- url: /
  description: The deployment serving this document
paths:
  /v1/jurisdictions:
    get:
      tags:
      - jurisdictions
      summary: List jurisdictions
      operationId: listJurisdictions
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JurisdictionPage'
          headers:
            Cache-Control: &id001
              $ref: '#/components/headers/Cache-Control'
            ETag: &id002
              $ref: '#/components/headers/ETag'
            RateLimit-Policy: &id003
              $ref: '#/components/headers/RateLimit-Policy'
            RateLimit: &id004
              $ref: '#/components/headers/RateLimit'
        '304':
          $ref: '#/components/responses/NotModified'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
      security:
      - ReadToken: []
      - IngestToken: []
  /v1/jurisdictions/{code}:
    get:
      tags:
      - jurisdictions
      summary: Get a jurisdiction with its registers and their freshness
      operationId: getJurisdiction
      security:
      - ReadToken: []
      - IngestToken: []
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
          description: 'Jurisdiction code, any case, `-` or `_`: `gb`, `US-NJ`, `ca_on`'
          examples:
          - ch
          title: Code
        description: 'Jurisdiction code, any case, `-` or `_`: `gb`, `US-NJ`, `ca_on`'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JurisdictionDetail'
          headers:
            Cache-Control: *id001
            ETag: *id002
            RateLimit-Policy: *id003
            RateLimit: *id004
        '304':
          $ref: '#/components/responses/NotModified'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /v1/jurisdictions/{code}/changes:
    get:
      tags:
      - changes
      summary: Rows added and removed across a jurisdiction's registers, oldest first
      description: Registers interleave by when their snapshots were recorded. Each register's first complete
        snapshot is its baseline, not a change.
      operationId: listJurisdictionChanges
      security:
      - ReadToken: []
      - IngestToken: []
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
          description: 'Jurisdiction code, any case, `-` or `_`: `gb`, `US-NJ`, `ca_on`'
          examples:
          - ch
          title: Code
        description: 'Jurisdiction code, any case, `-` or `_`: `gb`, `US-NJ`, `ca_on`'
      - name: since
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          description: Only changes recorded at or after this RFC 3339 time, with an offset (`2026-10-01T00:00:00Z`)
          title: Since
        description: Only changes recorded at or after this RFC 3339 time, with an offset (`2026-10-01T00:00:00Z`)
      - name: until
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          description: Only changes recorded before this time; after `since`
          title: Until
        description: Only changes recorded before this time; after `since`
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 1000
          minimum: 1
          description: Page size
          default: 100
          title: Limit
        description: Page size
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: '`next_cursor` from the previous page'
          title: Cursor
        description: '`next_cursor` from the previous page'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangePage'
          headers:
            Cache-Control: *id001
            ETag: *id002
            RateLimit-Policy: *id003
            RateLimit: *id004
        '304':
          $ref: '#/components/responses/NotModified'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '504':
          $ref: '#/components/responses/QueryTimeout'
  /v1/registers:
    get:
      tags:
      - registers
      summary: List registers
      operationId: listRegisters
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterPage'
          headers:
            Cache-Control: *id001
            ETag: *id002
            RateLimit-Policy: *id003
            RateLimit: *id004
        '304':
          $ref: '#/components/responses/NotModified'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
      security:
      - ReadToken: []
      - IngestToken: []
  /v1/registers/{slug}:
    get:
      tags:
      - registers
      summary: Get a register with its tables and freshness
      operationId: getRegister
      security:
      - ReadToken: []
      - IngestToken: []
      parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
          description: Register slug, e.g. `gb_ukgc`
          examples:
          - gb_ukgc
          title: Slug
        description: Register slug, e.g. `gb_ukgc`
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterDetail'
          headers:
            Cache-Control: *id001
            ETag: *id002
            RateLimit-Policy: *id003
            RateLimit: *id004
        '304':
          $ref: '#/components/responses/NotModified'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /v1/search:
    get:
      tags:
      - search
      summary: Find text in every register's current rows
      operationId: search
      security:
      - ReadToken: []
      - IngestToken: []
      parameters:
      - name: q
        in: query
        required: true
        schema:
          type: string
          minLength: 2
          maxLength: 200
          description: Case-insensitive substring
          title: Q
        description: Case-insensitive substring
      - name: jurisdiction
        in: query
        required: false
        schema:
          anyOf:
          - type: array
            items:
              type: string
          - type: 'null'
          description: Only these jurisdictions; repeat the parameter for several (`?jurisdiction=gb&jurisdiction=de`)
          title: Jurisdiction
        description: Only these jurisdictions; repeat the parameter for several (`?jurisdiction=gb&jurisdiction=de`)
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: Rows returned per table
          default: 20
          title: Limit
        description: Rows returned per table
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchHitPage'
          headers:
            Cache-Control: *id001
            ETag: *id002
            RateLimit-Policy: *id003
            RateLimit: *id004
        '304':
          $ref: '#/components/responses/NotModified'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '504':
          $ref: '#/components/responses/QueryTimeout'
...
  /registers:
    get:
      tags:
      - health
      summary: List Registers
      operationId: list_registers_registers_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  additionalProperties: true
                  type: object
                type: array
                title: Response List Registers Registers Get
        '400':
          description: A parameter or the path carried a NUL byte
      security: []
  /status:
    get:
      tags:
      - health
      summary: Status
      operationId: status_status_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Status Status Get
        '400':
          description: A parameter or the path carried a NUL byte
      security: []
  /ingest/{slug}:
    post:
      tags:
      - ingest
      summary: Trigger
      operationId: trigger_ingest__slug__post
      security:
      - IngestToken: []
      parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
          title: Slug
      - name: force
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Force
      - name: accept_count_delta
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Accept Count Delta
      - name: wait
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Wait
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Trigger Ingest  Slug  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /ingest/last:
    get:
      tags:
      - ingest
      summary: Last Run
      operationId: last_run_ingest_last_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Last Run Ingest Last Get
        '400':
          description: A parameter or the path carried a NUL byte
      security:
      - IngestToken: []
  /jurisdictions/{code}/ingest:
    post:
      tags:
      - ingest
      summary: Trigger Jurisdiction
      operationId: trigger_jurisdiction_jurisdictions__code__ingest_post
      security:
      - IngestToken: []
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
          title: Code
      - name: force
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Force
      - name: accept_count_delta
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Accept Count Delta
      - name: wait
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Wait
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Trigger Jurisdiction Jurisdictions  Code  Ingest Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /jurisdictions:
    get:
      tags:
      - read
      summary: List Jurisdictions
      operationId: list_jurisdictions_jurisdictions_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  additionalProperties: true
                  type: object
                type: array
                title: Response List Jurisdictions Jurisdictions Get
        '400':
          description: A parameter or the path carried a NUL byte
      security:
      - ReadToken: []
      - IngestToken: []
  /jurisdictions/{code}:
    get:
      tags:
      - read
      summary: Get Jurisdiction
      operationId: get_jurisdiction_jurisdictions__code__get
      security:
      - ReadToken: []
      - IngestToken: []
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
          title: Code
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Jurisdiction Jurisdictions  Code  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /jurisdictions/{code}/changes:
    get:
      tags:
      - read
      summary: Get Changes
      operationId: get_changes_jurisdictions__code__changes_get
      security:
      - ReadToken: []
      - IngestToken: []
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
          title: Code
      - name: since
        in: query
        required: true
        schema:
          type: string
          format: date-time
          description: ISO date or timestamp
          title: Since
        description: ISO date or timestamp
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          default: 200
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Changes Jurisdictions  Code  Changes Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /jurisdictions/{code}/{slug}/{table}:
    get:
      tags:
      - read
      summary: Get Rows
      description: 'Current rows. Any other query parameter filters a column by exact value,

        e.g. `/jurisdictions/gb/gb_ukgc/licences?status=Active&type=Remote`.'
      operationId: get_rows_jurisdictions__code___slug___table__get
      security:
      - ReadToken: []
      - IngestToken: []
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
          title: Code
      - name: slug
        in: path
        required: true
        schema:
          type: string
          title: Slug
      - name: table
        in: path
        required: true
        schema:
          type: string
          title: Table
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Q
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 100
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Rows Jurisdictions  Code   Slug   Table  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /search:
    get:
      tags:
      - read
      summary: Search
      operationId: search_search_get
      security:
      - ReadToken: []
      - IngestToken: []
      parameters:
      - name: q
        in: query
        required: true
        schema:
          type: string
          minLength: 2
          title: Q
      - name: jurisdiction
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Jurisdiction
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 20
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Search Search Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    JurisdictionPage:
      properties:
        data:
          items:
            $ref: '#/components/schemas/Jurisdiction'
          type: array
          title: Data
        pagination:
          $ref: '#/components/schemas/Pagination'
      type: object
      required:
      - data
      - pagination
      title: JurisdictionPage
      examples:
      - data:
        - *id009
        pagination:
          next_cursor: null
          has_more: false
          limit: 20
          total: 20
...
  securitySchemes:
    ReadToken:
      type: http
      description: '`READ_TOKEN`. Needed on read routes only when the server sets one; when it does not,
        reads are open. The ingest token reads too.'
      scheme: bearer
    IngestToken:
      type: http
      description: '`INGEST_TOKEN`. Starts and lists ingest runs, and reads. A read token here is a 403.'
      scheme: bearer
...
  headers:
    ETag:
      description: This representation's version (weak). Send it back in `If-None-Match` and the answer
        is a 304 with no body while the data is unchanged. New data shows within 30 seconds of its snapshot.
      schema:
        type: string
      example: W/"9f2c5d0e..."
    Cache-Control:
      description: '`public, max-age=300` when reads are open, `private, max-age=300` when the deployment
        requires a read token; `no-store` on ingest operations'
      schema:
        type: string
    RateLimit-Policy:
      description: 'The quota this request counts against (draft-ietf-httpapi-ratelimit-headers): `"read";q=600;w=60`
        is 600 requests per 60 seconds. Per bearer token, or per client address without one. Defaults:
        `read` 600, `search` (search and domains) 60, `ingest` (starting runs) 10 a minute.'
      schema:
        type: string
      example: '"search";q=60;w=60'
...